+++ src/Gnugat/Phpixture/Repository.php
@@ -2,6 +2,10 @@
 
 namespace Gnugat\Phpixture;
 
+use Gnugat\Phpixture\ValueSwitcher\KeepAsIsStrategy;
+use Gnugat\Phpixture\ValueSwitcher\ToOneRelationshipStrategy;
+use Gnugat\Phpixture\ValueSwitcher\ValueSwitcher;
+
 class Repository
 {
     /**
@@ -10,11 +14,19 @@ class Repository
     private $fixtures;
 
     /**
+     * @var ValueSwitcher
+     */
+    private $valueSwitcher;
+
+    /**
      * @param array $fixtures
      */
     public function __construct(array $fixtures)
     {
         $this->fixtures = $fixtures;
+        $this->valueSwitcher = new ValueSwitcher();
+        $this->valueSwitcher->add(new ToOneRelationshipStrategy($fixtures));
+        $this->valueSwitcher->add(new KeepAsIsStrategy());
     }
 
     /**
@@ -26,9 +38,19 @@ class Repository
     {
         $fixtures = array();
         foreach ($this->fixtures[$name] as $fixture) {
-            $fixtures[] = $fixture;
+            $fixtures[] = $this->handleRelationships($fixture);
         }
 
         return $fixtures;
+    }
+
+    private function handleRelationships(array $fields)
+    {
+        $replacedFields = array();
+        foreach ($fields as $name => $value) {
+            $replacedFields[$name] = $this->valueSwitcher->switchValue($value);
+        }
+
+        return $replacedFields;
     }
 }