+++ src/Gnugat/Phpixture/ValueSwitcher/ValueSwitcherStrategy.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Gnugat\Phpixture\ValueSwitcher;
+
+interface ValueSwitcherStrategy
+{
+    /**
+     * @param mixed $value
+     *
+     * @return bool
+     */
+    public function supports($value);
+
+    /**
+     * @param mixed $value
+     *
+     * @return mixed
+     */
+    public function switchValue($value);
+}