+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.2.0: ToOne relationship management
+
+* Created KeepAsIsStrategy
+* Created ToOneRelationshipStrategy
+* Created ValueSwitcher
+
 ## 0.1.0: Simple fixtures
 
 * Created Repository#findAll